requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
WTForms==3.0.1
Flask-WTF==1.1.1
python-barcode>=0.15.1
Pillow>=10.3.0
gunicorn==22.0.0